feat: 필터링된 참여자들을 구분하기 위한 overlay 구현
